+++ CHANGELOG.md
@@ -8,7 +8,7 @@ _(none)_
 ---
 
 ## 2.0.1 (2020-08-22)
-_(none)_
+* Patch: Fix bin definition in package
 
 ## 2.0.0 (2020-08-22)
 * Breaking: Improved Scenario Outline syntax