+++ CHANGELOG.md
@@ -2,12 +2,15 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+_(none)_
+
+---
+
+## 2.1.0 (2020-08-23)
 * Feature: Improved `scenarioOutline` syntax
 * Feature: Transform uses improved `scenarioOutline` syntax
 * Feature: Support `gherkin.skip` and `gherkin.only`
 * Feature: Even more improved Scenario Outline syntax
-
----
 
 ## 2.0.1 (2020-08-22)
 * Patch: Fix bin definition in package